Get rid of the concept of expiring sessions at a particular point in time
Sessions expire when the user logs out, but not after a while. Unintuitive,
but it _is_ a best practice. Users are not used to sessions expiring, and
the rest of the application won't be coded to the standard of expecting
the session to expire mid-operation. There's security benefits as well, as
a user who is trained on the session expiring will more eagerly provide
credentials in a malicious page.